implementations - task 13

+++ src/main/java/com/adamludzia/db/impl/InMemoryDatabase.java
@@ -14,11 +14,11 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class InMemoryDatabase implements Database {
 
-    private final Map<Integer, Invoice> invoices = new HashMap<>();
-    private int nextId = 1;
+    private final Map<Long, Invoice> invoices = new HashMap<>();
+    private long nextId = 1;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         invoice.setId(nextId);
         invoices.put(nextId, invoice);
 
@@ -26,7 +26,7 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         return Optional.ofNullable(invoices.get(id));
     }
 
@@ -36,13 +36,13 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         updatedInvoice.setId(id);
         return Optional.ofNullable(invoices.put(id, updatedInvoice));
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         return Optional.ofNullable(invoices.remove(id));
     }
 }
